Backend_tshirtop1/productManagement/productSearchGraphQL/requirements.txt
Flask
Flask-GraphQL
graphene
graphql-core
mysql-connector-python
python-dotenv
SQLAlchemy
